+++ ESI.NET/Enumerations/GrantType.cs
@@ -3,6 +3,7 @@ using System.Runtime.Serialization;
 
 namespace ESI.NET.Enumerations
 {
+    [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
     public enum GrantType
     {
         [EnumMember(Value="authorization_code")] /**/ AuthorizationCode,